Refactored cli into client directory
