+++ novelsave/__main__.py
@@ -1,4 +1,4 @@
-from .cli import main
+from novelsave.client.cli import main
 
 
 if __name__ == "__main__":
