+++ novelsave/client/cli/controllers/_url.py
@@ -3,7 +3,7 @@ import sys
 from dependency_injector.wiring import inject, Provide
 from loguru import logger
 
-from novelsave.cli.helpers import get_novel
+from novelsave.client.cli.helpers import get_novel
 from novelsave.containers import Application
 from novelsave.core.services import BaseNovelService
