+++ novelsave/client/cli/main.py
@@ -8,12 +8,12 @@ from tqdm import tqdm
 
 from . import controllers, helpers, groups, events
 from .events import update_check_event
-from .. import __version__
-from ..containers import Application
-from ..exceptions import NSError
-from ..migrations import commands as migration_commands
-from ..settings import config, DATABASE_URL, LOGGER_CONFIG
-from ..utils.helpers import config_helper
+from novelsave import __version__
+from novelsave.containers import Application
+from novelsave.exceptions import NSError
+from novelsave.migrations import commands as migration_commands
+from novelsave.settings import config, DATABASE_URL, LOGGER_CONFIG
+from novelsave.utils.helpers import config_helper
 
 
 def inject_dependencies():
